review bugs fix

+++ Tracker/Store/TrackerRecordStore.swift
@@ -4,38 +4,30 @@ import UIKit
 import CoreData
 
 protocol TrackerRecordStoreProtocol: AnyObject {
-    func addNewRecord(tracker: Tracker, trackerRecord: TrackerRecord) throws
+    func addNewRecord(trackerCoreData: TrackerCoreData, trackerRecord: TrackerRecord) throws
     func getAllRecords() -> [TrackerRecord]?
     func removeRecord(tracker: Tracker, trackerRecord: TrackerRecord) throws
 }
 
 final class TrackerRecordStore {
     private let context: NSManagedObjectContext
-    private lazy var trackerStore: TrackerStoreGetTrackerCoreDataForIdProtocol = TrackerStore(context: self.context)
-    
-    convenience init() {
-        let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-        self.init(context: context)
-    }
+    private lazy var trackerStore: TrackerStore = TrackerStore(context: self.context)
     
     init(context: NSManagedObjectContext) {
         self.context = context
     }
-    
 }
 
 //MARK: TrackerRecordStoreProtocol func
+
 extension TrackerRecordStore: TrackerRecordStoreProtocol {
-    func addNewRecord(tracker: Tracker, trackerRecord: TrackerRecord) throws {
-        
+    func addNewRecord(trackerCoreData: TrackerCoreData, trackerRecord: TrackerRecord) throws {
         let trackerRecordCoreData = TrackerRecordCoreData(context: context)
         
-        let trackerCoreData = trackerStore.getTrackerCoreDataForId(id: tracker.id)
         trackerRecordCoreData.dateOfCompletion = trackerRecord.dateOfCompletion
         trackerRecordCoreData.tracker = trackerCoreData
-        do {
-            try context.save()
-        }
+        try? context.save()
+        
     }
     
     func getAllRecords() -> [TrackerRecord]? {
@@ -60,7 +52,8 @@ extension TrackerRecordStore: TrackerRecordStoreProtocol {
         let trackerRecordsCoreData = try? context.fetch(fetchRequest)
         
         if let currentRecord = trackerRecordsCoreData?.first(where: {
-            $0.tracker?.id == tracker.id && Calendar.current.isDate($0.dateOfCompletion!, inSameDayAs: trackerRecord.dateOfCompletion)
+            guard let dateOfCompletion = $0.dateOfCompletion else {return false}
+            return $0.tracker?.id == tracker.id && Calendar.current.isDate(dateOfCompletion, inSameDayAs: trackerRecord.dateOfCompletion)
         }) {
             context.delete(currentRecord)
             try? context.save()
